Support systemd socket activation in mbserver.
systemd can activate services when clients attempt to connect to the
sockets or file descriptors over which they communicate.  This leads
to fewer manually-managed dependencies in startup, making it possible
to minimize system boot time and restart crashed services.  Membroker
and his clients are not able to deal with crashing and restarting at
present, but this makes it more of a possibility.  At the very least,
we want the simplication this provides for boot sequence configuration.

If inst_systemdsystemunitdir is nonempty, then link with libsystemd to
get our listen socket passed to us, and install service configuration
files to make membroker a socket-activated systemd service.

Refactor initialization to make it possible to pass in a pre-connected
socket.  The unit test and standalone operation of mbserver remain
unchanged.