mbserver: find the actual pid of the client for more useful debug reporting
The id is passed in by the client, and is not guaranteed to be unique.
If the id is not the pid, we get "unknown" as the host process name
in debug, making it very difficult to figure out what is going on.

Ask the unix domain socket for the pid of the process on the other end,
and store this in our config.  Use this to find the name of the other
process.

I would like to change the protocol to force the pid to part of the
infrastructure, and ids to be an optional add-on.  The ids are necessary
for supporting more than one distinct membroker connection per process,
so we can't rely on pid alone.  Also, the system doesn't give us tids,
and the thread creating the connect may not be the one servicing it...
Should we allow user-supplied names?