Fixed spurious failure in testDumpDebug.
2nd request was not always getting rejected since with the right sequence
of thread execution it was sometimes possible for membroker to complete
the first request before reading the 2nd. Changed the request to a reserve,
which triggers 2 sequential queries to the source client and ensures that
the 2nd request will be read between them.

Also moved the debug test to the middle of the queries so that it exercises
more of the debug code.