mbserver: factor out cmdline reading code, fix minor edge case bug
Coverity complained about the possibility of read() failing for some
reason other than EINTR, causing us to write to array index -1, which
would corrupt something on the stack.

In reality, since this is a proc file that we have already successfully
opened, we're not going to fail to read it; the kernel just writes some
data back to the buffer we passed in.

But, clean code is happy code, and the large amount of logic to read
this data file in the middle of initializing a client connection is
poor form... so, move the code to a helper function and fix this edge
case.

CID 28389  Negative array index write

EV 380672  900FE w/ 4GB DIMM

Review-link: https://reviews.lpdev.prtdev.lexmark.com/r/44556/
Signed-off-by: Scott Arrington <sarringt@lexmark.com>