mbserver: clarify initialization diagnostic messages
The "Initialized membroker with %d pages" message at start up is
misleading.  It is intended to be a diagnostic for gmalloc clients
who get initial server pages via environment variable.  The message
comes out too early to be useful as a diagnostic for a situation in
which you explicitly set the server's initial pages with
mb_server_set_pages() -- so you see that it was initialized with
zero pages, and then it has some.  Wha?

So, hush that diagnostic, and add a different one.

EV 380672  900FE w/ 4GB DIMM

Review-link: https://reviews.lpdev.prtdev.lexmark.com/r/44560/
Signed-off-by: Scott Arrington <sarringt@lexmark.com>