Fixed bug when a source client terminates
This could leave membroker with more outstanding pages than it can back.

Converted server->pages to an int so it can accurately reflect
the negative balance of pages, and converted most other page counters
in mbserver to ints to eliminate type mismatches.

Added a check in testIoErrors unit test to verify that membroker
will not disburse pages to clients if it has a negative page balance.